Rename `laundry-washer-controls-server` to `CodegenIntegration`
