+++ src/app/clusters/laundry-washer-controls-server/CodegenIntegration.cpp
@@ -20,7 +20,7 @@
 #include <app/util/config.h>
 
 #include "laundry-washer-controls-delegate.h"
-#include "laundry-washer-controls-server.h"
+#include "CodegenIntegration.h"
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app-common/zap-generated/callback.h>
 #include <app-common/zap-generated/cluster-objects.h>
