+++ src/app/clusters/laundry-washer-controls-server/app_config_dependent_sources.cmake
@@ -17,6 +17,6 @@ TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
     "${CLUSTER_DIR}/laundry-washer-controls-delegate.h"
-    "${CLUSTER_DIR}/laundry-washer-controls-server.cpp"
-    "${CLUSTER_DIR}/laundry-washer-controls-server.h"
+    "${CLUSTER_DIR}/CodegenIntegration.cpp"
+    "${CLUSTER_DIR}/CodegenIntegration.h"
 )
